plugins/bibliogon-plugin-kdp/pyproject.toml
[tool.poetry]
name = "bibliogon-plugin-kdp"
version = "1.0.0"
description = "Amazon KDP plugin for Bibliogon - metadata export, cover validation, preview"
authors = ["astrapi69"]
license = "Proprietary"
readme = "README.md"
packages = [{include = "bibliogon_kdp"}]

[tool.poetry.dependencies]
python = "^3.11"
pluginforge = {path = "../../pluginforge", develop = true}
fastapi = "^0.115.0"
pyyaml = "^6.0"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[tool.poetry.plugins."bibliogon.plugins"]
kdp = "bibliogon_kdp.plugin:KdpPlugin"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
